move construction logic out of parse.go into construct.go
